Add a generic `Executor` trait
This commit adds a new trait to the `future` module, `Executor`. This trait has
only one method:

    trait Execute<F: Future<Item = (), Error = ()>> {
        fn execute(&self, f: F) -> Result<(), ExecuteError<F>>;
    }

The purpose of this trait is to unify the various executors found throughout the
ecosystem. Crates which require the ability to spawn futures will now have the
option ot operate generically over all `Executor` instances instead of a
particular executor.

A `Future::background` method was also added to ergonomically pass a future to
an executor and run it in the background. The two `oneshot` modules (sync and
unsync) also grew `spawn` and `spawn_fn` functions which will spawn a future
onto an executor, returning a handle to the resulting future. This can be used
to spawn work onto a specific executor and still get notified when the work
itself is completed.

Finally, an implementation of the `Executor` trait was added to `CpuPool`. Due
to the differences in unwinding/panic semantics, though, the `CpuPool::spawn`
method which previously existed was not deprecated.

Closes #313